Explicitly check if tmpfile is nil before calling close
Fixes #246

+++ lib/vagrant-berkshelf/action/upload.rb
@@ -57,7 +57,7 @@ module VagrantPlugins
             tmpfile.rewind
             yield tmpfile.path
           ensure
-            if defined?(tmpfile)
+            if defined?(tmpfile) && !tmpfile.nil?
               tmpfile.close
               tmpfile.unlink
             end
